+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 # 1.6.5
 
+- [Feature] Infrared Screen with old style
 - [Feature] Flipper Busy Dialog with remote
 - [Feature] Infrared Editor
 - [Feature] Scroll on Infrared Screen